+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -70,7 +70,8 @@ namespace Smartstore.Core.Data.Migrations
                 "Media.Manufacturer.ImageAlternateTextFormat",
                 "Media.Product.ImageAlternateTextFormat",
                 "Common.DecreaseValue",
-                "Common.IncreaseValue");
+                "Common.IncreaseValue",
+                "Admin.Orders.Address.EditAddress");
 
             builder.AddOrUpdate("ShoppingCart.DiscountCouponCode.Removed", "The discount code has been removed", "Der Rabattcode wurde entfernt");
             builder.AddOrUpdate("ShoppingCart.GiftCardCouponCode.Removed", "The gift card code has been removed", "Der Geschenkgutschein wurde entfernt");