+++ src/Smartstore.Web/Areas/Admin/Models/Orders/OrderAddressModel.cs
@@ -6,6 +6,7 @@ namespace Smartstore.Admin.Models.Orders
     public class OrderAddressModel : ModelBase
     {
         public int OrderId { get; set; }
+        public string Username { get; set; }
 
         [UIHint("Address")]
         public AddressModel Address { get; set; } = new();